test: split logsQueryBuilder.spec.js into 3 files and rebalance CI shards
Split the 82-test monolithic logsQueryBuilder.spec.js (1,968 lines) into 3
parallel-friendly files: logsQueryBuilder-chart (28 tests), logsQueryBuilder-filters
(26 tests), logsQueryBuilder-editor (23 tests). Redistributes all Logs tests
from 2 uneven shards (Logs-Core 175, Logs-Queries 57) into 3 balanced shards
(Logs-Builder ~102, Logs-Core ~92, Logs-Features ~77), cutting shard run time
from ~28 min to ~10-12 min each.
